agentic_jobs/extract_benchmarking_agent_job/requirements.txt
fastapi
uvicorn[standard]
google-adk
google-cloud-firestore
google-cloud-storage
google-cloud-secret-manager
google-genai
python-dotenv
google-auth
PyMuPDF
requests
beautifulsoup4
tavily-python
aiohttp